refactor(error): split tests_variants.rs (#1266)
tests_variants.rs 416 -> 255 LOC; new tests_variants_extra.rs (~167 LOC).
